+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * [#6](https://github.com/TinkerDev/grape-swagger-rails/pull/6): Fix: support multiple predefined headers - [@Tyr0](https://github.com/tyr0).
 * Your contribution here.
 
+* Added a new :custom_css_file option to have default index.html.erb render a stylesheet link for custom theming - [@alexagranov](https://github.com/alexagranov).
+
 ### 0.1.0 (February 5, 2015)
 
 * [41](https://github.com/BrandyMint/grape-swagger-rails/pull/41): Compatibility with grape-swagger 0.8.0 and 0.9.0 - [@dblock](https://github.com/dblock).